Chapter 15: A Better Application Structure (v0.15)
